Move example instances to own file
When more factor and variable classes are defined in later PR, they will get
their own proper files.
