+++ wave_optimization/include/wave/optimization/factor_graph/FactorVariable.hpp
@@ -59,44 +59,6 @@ class FactorVariable : public FactorVariableBase {
     Eigen::Map<const Eigen::Matrix<double, Size, 1>> map;
 };
 
-/**
- * Specialized variable representing a 2D pose.
- *
- * In this example, references are assigned to parts of the underlying data
- * vector, allowing factor functions to operate on clearly named parameters.
- */
-struct Pose2DVar : public FactorVariable<3> {
-    Pose2DVar() = default;
-
-    // Use base class constructor
-    // We can't inherit constructors due to bug in gcc
-    // (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=67054)
-    explicit Pose2DVar(const double *d) : FactorVariable<3>{d} {}
-
-    using Vec1 = Eigen::Matrix<double, 1, 1>;
-
-    Eigen::Ref<const Vec2> position{map.head<2>()};
-    Eigen::Ref<const Vec1> orientation{map.tail<1>()};
-};
-
-/**
- * Specialized variable representing a 2D landmark position.
- *
- * In this example, references are assigned to parts of the underlying data
- * vector, allowing factor functions to operate on clearly named parameters.
- */
-struct Landmark2DVar : public FactorVariable<2> {
-    Landmark2DVar() = default;
-
-    // Use base class constructor
-    // We can't inherit constructors due to bug in gcc
-    // (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=67054)
-    explicit Landmark2DVar(const double *d) : FactorVariable<2>{d} {}
-
-    Eigen::Map<const Vec2> position{map.data()};
-};
-
-
 /** @} group optimization */
 }  // namespace wave
